Fixed wrong ID type in ProductRepository

+++ src/main/java/com/jseric/simple_product_rest/repository/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends SoftDeleteRepository<Product, Integer> {
+public interface ProductRepository extends SoftDeleteRepository<Product, Long> {
 
     @Query("SELECT COUNT(p.id) FROM Product p WHERE p.code = ?1 AND p.deleted IS NULL")
     long countByCode(final String code);
